newclub description field fixes and deleted unused code
